pyproject.toml
[project]
name = "chapter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "docling-mcp>=1.3.3",
]
